+++ component_catalog/forms.py
@@ -161,7 +161,7 @@ class ComponentForm(
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.fields["owner"].dataspace = self.dataspace
+        self.fields["owner"].user = self.user
 
     def clean_packages_ids(self):
         packages_ids = self.cleaned_data.get("packages_ids")